Create database and table and insert log data
When run, the dash_sqlite.py file creates a database and establishes a connection. It then creates a table and using the confluent kafka consumer it polls the topic and inserts the logs into the table. When the log is the user information the whole table is queried and turned into a csv. the table is then delete and a new one created. The csv will be pushed to an s3 bucket when the name of the bucket is known.