Create the basic framework for the Daily Report
Using a html file the template for the daily report is created. This is then loaded into the python file using jinja2 where the relevant variables are added. Currently, placeholders are showing where these graphs and metrics would go. Using pdfkit and wkhtmltopdf this text is converted into pdf format using the styling css in the same folder.